pyproject.toml
# pyproject.toml

[tool.setuptools]
package-dir = {"src" = "dydx4"}

[tool.poetry]
name = "dydx-v4-python"
version = "0.0.1"
description = "dYdX Chain Client"
authors = ["amiwrpremium@gmail.com"]
license = "AGPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
Cython = "^0.29.0"
aiohttp = "^3.8.1"
bech32 = "^1.2.0"
bip_utils = "^2.7.0"
brownie = "^0.5.0"
cytoolz = "^0.12.1"
dateparser = "^1.0.0"
ecdsa = "^0.18.0"
eth_keys = "^0.4.0"
eth-account = "^0.9.0"
protobuf = "^4.23"
grpcio-tools = "^1.56"
grpcio = "^1.56"
mpmath = "^1.3.0"
requests = "^2.31.0"
six = "^1.16"
sympy = "^1.12.0"
web3 = "^6.5.0"
websocket_client = "^1.6.1"
v4-proto = "^2.0.0"
blspy = "^2.0.2"

[tool.poetry.dev-dependencies]
wheel = "^0.35.1"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pytype = "^2023.12.8"
pydocstyle = "^6.3.0"
black = "^23.12.0"
types-python-dateutil = "^2.8.19.14"
types-requests = "^2.31.0.10"
types-protobuf = "^4.24.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
